updated contexts.py to allow display of basket contents + basic styling of basket.html
