--- basket/context.py
@@ -1,36 +1,0 @@
-from decimal import Decimal
-from django.conf import settings
-from django.shortcuts import get_object_or_404
-from products.models import Product
-
-""" 
-Context processor for the basket to allow basket
-to be available across all apps.
-"""
-
-def basket_contents(request):
-
-    basket_items = []
-    total = 0
-    product_count = 0
-
-    if total < settings.FREE_DELIVERY_THRESHOLD:
-        delivery = total * Decimal(settings.STANDARD_DELIVERY_PERCENTAGE / 100)
-        free_delivery_delta = settings.FREE_DELIVERY_THRESHOLD - total
-    else:
-        delivery = 0
-        free_delivery_delta = 0
-
-    grand_total = delivery + total
-
-    context = {
-        "basket_items": basket_items,
-        "total": total,
-        "product_count": product_count,
-        "delivery": delivery,
-        "free_delivery_delta": free_delivery_delta,
-        "free_delivery_threshold": settings.FREE_DELIVERY_THRESHOLD,
-        "grand_total": grand_total,
-    }
-
-    return context
