+++ static/css/base.css
@@ -270,3 +270,13 @@ nav {
     max-height:100%;
     object-fit: cover;
 }
+
+/* Styling for basket */
+
+.basket-image {
+    height: 100px;
+    width: 100px;
+    display: flex;
+    align-items: center;
+    object-fit: contain;
+}
