+++ templates/includes/navbar.html
@@ -19,7 +19,7 @@
                     <p id="cart-total" class="m-0">
                     <i class="nav-icon fas fa-shopping-cart"></i>
                         {% if grand_total %}
-                            ${{ grand_total|floatformat:2 }}
+                            £{{ grand_total|floatformat:2 }}
                         {% else %}
                             £0.00
                         {% endif %}
@@ -90,7 +90,7 @@
                     <p id="cart-total" class="m-0">
                     <i class="nav-icon fas fa-shopping-cart"></i>
                         {% if grand_total %}
-                            ${{ grand_total|floatformat:2 }}
+                            £{{ grand_total|floatformat:2 }}
                         {% else %}
                             £0.00
                         {% endif %}
@@ -100,4 +100,5 @@
         </ul>
     </div>            
 </nav>
+
